# contracts/vuelos-ecommerce-openapi.yaml
...
openapi: 3.0.0
info:
  description: >-
    Contrato del sistema de Vuelos tal como está implementado: núcleo GDS (búsqueda, retención de cupos, reservas, billetes, posventa,
    check-in, webhooks), e-commerce (identidad, catálogo, ofertas, pagos simulados, órdenes) y administración (órdenes, vuelos, usuarios,
    dashboard, observabilidad).


    Convenciones comunes: prefijo /api/v1; autenticación Bearer JWT (HS256) y `ownerId` siempre igual al `sub` verificado; errores
    `application/problem+json` con el campo `code`; las escrituras críticas exigen `Idempotency-Key` (UUID); los importes en el borde de la
    API son cadenas decimales y en el sistema son enteros en unidades menores.
  title: RAM Alliance · API de Vuelos (sistema completo)
  version: '1.0'
servers:
  - description: Producción (Railway)
    url: https://proyect-flights-rda1-plantilla-production.up.railway.app/api/v1
  - description: Desarrollo local
    url: http://localhost:3000/api/v1
...
paths:
...
  /admin/ordenes:
    get:
      description: Más recientes primero, paginado por cursor; filtros por estado, número, PNR y fechas.
      operationId: AdminController_ordenes
      parameters:
        - in: query
          name: limit
          required: false
          schema:
            default: 25
            maximum: 100
            minimum: 1
            type: number
        - description: Cursor opaco de la página anterior (nextCursor).
          in: query
          name: cursor
          required: false
          schema:
            type: string
        - in: query
          name: estado
          required: false
          schema:
            enum:
              - PENDIENTE_PAGO
              - PAGO_EN_VERIFICACION
              - PAGADA
              - EMITIDA
              - FALLIDA_COMPENSADA
              - MODIFICADA
              - DEVOLUCION_EN_CURSO
              - REEMBOLSADA
              - EN_VIAJE
              - COMPLETADA
              - EXPIRADA
            type: string
        - in: query
          name: numero
          required: false
          schema:
            example: ORD-7K3M9PQ2XA
            type: string
        - in: query
          name: pnr
          required: false
          schema:
            example: ABC123
            type: string
        - description: Creadas desde este día (UTC).
          in: query
          name: desde
          required: false
          schema:
            format: date
            type: string
        - description: Creadas hasta este día (UTC, inclusive).
          in: query
          name: hasta
          required: false
          schema:
            format: date
            type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminOrdenesPaginaDto'
          description: ''
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetailsDto'
          description: Petición inválida
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetailsDto'
          description: No autenticado
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetailsDto'
          description: Prohibido
      security:
        - bearer: []
      summary: Listar órdenes de todos los clientes (ADMIN)
      tags:
        - E-commerce · Back-office
  /admin/ordenes/{numero}:
    get:
      description: Incluye el contacto del comprador.
      operationId: AdminController_orden
      parameters:
        - description: Número de orden
          in: path
          name: numero
          required: true
          schema:
            example: ORD-7K3M9PQ2XA
            type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminOrdenViewDto'
          description: ''
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetailsDto'
          description: Petición inválida
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetailsDto'
          description: No autenticado
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetailsDto'
          description: Prohibido
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetailsDto'
          description: No encontrado
      security:
        - bearer: []
      summary: Detalle completo de una orden (ADMIN)
      tags:
        - E-commerce · Back-office
...
  /admin/vuelos:
    get:
      description: Por salida ascendente, desde hoy si no se indica fecha; filtros por ruta, día y número de vuelo.
      operationId: AdminController_vuelos
      parameters:
        - in: query
          name: limit
          required: false
          schema:
            default: 25
            maximum: 100
            minimum: 1
            type: number
        - description: Cursor opaco de la página anterior (nextCursor).
          in: query
          name: cursor
          required: false
          schema:
            type: string
        - in: query
          name: origen
          required: false
          schema:
            example: BOG
            type: string
        - in: query
          name: destino
          required: false
          schema:
            example: SCL
            type: string
        - description: 'Día de salida (UTC). Sin fecha: desde hoy.'
          in: query
          name: fecha
          required: false
          schema:
            format: date
            type: string
        - in: query
          name: vuelo
          required: false
          schema:
            example: LA800
            type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminVuelosPaginaDto'
          description: ''
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetailsDto'
          description: Petición inválida
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetailsDto'
          description: No autenticado
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetailsDto'
          description: Prohibido
      security:
        - bearer: []
      summary: Listar vuelos con su inventario (ADMIN)
      tags:
        - E-commerce · Back-office
    post:
      description: >-
        Agrega un vuelo al calendario: todos sus asientos quedan disponibles y aparece en la búsqueda. El código lleva la aerolínea en sus
        dos primeros caracteres; origen y destino deben existir en el catálogo; la salida debe ser futura. 409 si ese código ya sale a esa
        hora.
      operationId: AdminController_crearVuelo
      parameters: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminCrearVueloDto'
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminVueloViewDto'
          description: ''
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetailsDto'
          description: Petición inválida
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetailsDto'
          description: No autenticado
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetailsDto'
          description: Prohibido
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetailsDto'
          description: Conflicto
      security:
        - bearer: []
      summary: Crear un vuelo (ADMIN)
      tags:
        - E-commerce · Back-office
  /admin/vuelos/{vueloId}:
    delete:
      description: >-
        Borra un vuelo que nunca tuvo retenciones, reservas ni asientos asignados. Si tiene historial responde 409 FLIGHT_IN_USE: cancélalo
        en su lugar (cierra la venta y reembolsa las reservas).
      operationId: AdminController_eliminarVuelo
      parameters:
        - in: path
          name: vueloId
          required: true
          schema:
            format: uuid
            type: string
      responses:
        '204':
          description: Eliminado
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetailsDto'
          description: Petición inválida
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetailsDto'
          description: No autenticado
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetailsDto'
          description: Prohibido
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetailsDto'
          description: No encontrado
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetailsDto'
          description: Conflicto
      security:
        - bearer: []
      summary: Eliminar un vuelo (ADMIN)
      tags:
        - E-commerce · Back-office
    patch:
      description: >-
        Cambia aerolínea, tarifa base, duración (recalcula la llegada) o capacidad. Solo vuelos programados que no han salido. La capacidad
        no puede bajar de los asientos ya vendidos ni dejar asientos numerados fuera de la cabina. Para mover la salida usa reprogramar.
      operationId: AdminController_actualizarVuelo
      parameters:
        - in: path
          name: vueloId
          required: true
          schema:
            format: uuid
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminActualizarVueloDto'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminVueloViewDto'
          description: ''
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetailsDto'
          description: Petición inválida
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetailsDto'
          description: No autenticado
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetailsDto'
          description: Prohibido
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetailsDto'
          description: No encontrado
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetailsDto'
          description: Conflicto
      security:
        - bearer: []
      summary: Editar un vuelo (ADMIN)
      tags:
        - E-commerce · Back-office
  /admin/vuelos/{vueloId}/asientos:
    get:
      description: >-
        Mapa de la cabina con la disponibilidad de cada asiento y la lista de reservados con su código de reserva y orden. No muestra
        nombres de pasajeros.
      operationId: AdminController_asientos
      parameters:
        - in: path
          name: vueloId
          required: true
          schema:
            format: uuid
            type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminAsientosVueloDto'
          description: ''
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetailsDto'
          description: Petición inválida
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetailsDto'
          description: No autenticado
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetailsDto'
          description: Prohibido
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetailsDto'
          description: No encontrado
      security:
        - bearer: []
      summary: Asientos reservados de un vuelo (ADMIN)
      tags:
        - E-commerce · Back-office
  /admin/vuelos/{vueloId}/cancelar:
    post:
      description: >-
        La aerolínea cancela el vuelo: se cierra a la venta, cada reserva confirmada que lo tenía se cancela con reembolso total (sin
        penalidad, sea cual sea la tarifa) y se emiten los eventos de webhook flight.cancelled y booking.cancelled. El reembolso llega a la
        pasarela y a la orden por esos eventos.
      operationId: AdminController_cancelarVuelo
      parameters:
        - in: path
          name: vueloId
          required: true
          schema:
            format: uuid
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminCancelarVueloDto'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminAccionVueloViewDto'
          description: ''
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetailsDto'
          description: Petición inválida
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetailsDto'
          description: No autenticado
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetailsDto'
          description: Prohibido
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetailsDto'
          description: No encontrado
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetailsDto'
          description: Conflicto
      security:
        - bearer: []
      summary: Cancelar un vuelo (ADMIN)
      tags:
        - E-commerce · Back-office
  /admin/vuelos/{vueloId}/reprogramar:
    post:
      description: >-
        La aerolínea mueve la salida (la duración se conserva): las reservas confirmadas con ese vuelo se actualizan y se emiten los eventos
        de webhook flight.schedule_changed y booking.changed. Responde 409 si otro vuelo con el mismo número ya sale a esa hora.
      operationId: AdminController_reprogramarVuelo
      parameters:
        - in: path
          name: vueloId
          required: true
          schema:
            format: uuid
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminReprogramarVueloDto'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminAccionVueloViewDto'
          description: ''
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetailsDto'
          description: Petición inválida
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetailsDto'
          description: No autenticado
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetailsDto'
          description: Prohibido
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetailsDto'
          description: No encontrado
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetailsDto'
          description: Conflicto
      security:
        - bearer: []
      summary: Reprogramar un vuelo (ADMIN)
      tags:
        - E-commerce · Back-office
...
  /ordenes:
    get:
      description: 'RF-ORD-010 / RF-CHK-001: acceso público para quien compró sin cuenta. No devuelve datos de contacto y limita los intentos por IP.'
      operationId: OrdenesController_recuperar
      parameters:
        - in: query
          name: numero
          required: false
          schema:
            example: ORD-7K3M9PQ2XA
            type: string
        - description: Código de reserva (PNR)
          in: query
          name: pnr
          required: false
          schema:
            example: ABC123
            type: string
        - description: Apellido de cualquier pasajero.
          in: query
          name: apellido
          required: true
          schema:
            example: Tapia
            type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrdenViewDto'
          description: ''
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetailsDto'
          description: Petición inválida
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetailsDto'
          description: No encontrado
        '429':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetailsDto'
          description: Demasiadas peticiones
      summary: Recuperar un viaje con número de orden o PNR y apellido
      tags:
        - E-commerce · Compra y órdenes
  /ordenes/{numero}:
    get:
      operationId: OrdenesController_obtener
      parameters:
        - description: Número de orden
          in: path
          name: numero
          required: true
          schema:
            example: ORD-7K3M9PQ2XA
            type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrdenViewDto'
          description: ''
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetailsDto'
          description: Petición inválida
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetailsDto'
          description: No autenticado
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetailsDto'
          description: No encontrado
      security:
        - bearer: []
      summary: Consultar una orden propia
      tags:
        - E-commerce · Compra y órdenes
...
components:
  schemas:
...
    AdminActualizarVueloDto:
      properties:
        aerolinea:
          example: LATAM Airlines
          type: string
        capacidadTotal:
          description: No puede ser menor que los asientos ya vendidos ni dejar asientos numerados fuera de la cabina.
          maximum: 400
          minimum: 1
          type: number
        duracionMinutos:
          description: Recalcula la llegada conservando la salida.
          maximum: 1500
          minimum: 20
          type: number
        precioBaseUsd:
          description: Tarifa base por adulto en USD. No afecta a las reservas ya hechas.
          example: 340
          type: number
      type: object
...
    AdminAsientosVueloDto:
      properties:
        capacidadTotal:
          type: number
        codigoVuelo:
          example: LA800
          type: string
        destino:
          example: SCL
          type: string
        filas:
          description: Mapa completo de la cabina con la disponibilidad de cada asiento.
          items:
            $ref: '#/components/schemas/AdminFilaMapaDto'
          type: array
        origen:
          example: BOG
          type: string
        reservados:
          description: Asientos reservados, por número.
          items:
            $ref: '#/components/schemas/AdminAsientoReservadoDto'
          type: array
        salida:
          type: string
        vueloId:
          format: uuid
          type: string
      required:
        - vueloId
        - codigoVuelo
        - origen
        - destino
        - salida
        - capacidadTotal
        - reservados
        - filas
      type: object
...
    AdminCancelarVueloDto:
      properties:
        motivo:
          example: Falla técnica de la aeronave
          type: string
      type: object
...
    AdminCrearVueloDto:
      properties:
        aerolinea:
          example: LATAM Airlines
          type: string
        capacidad:
          default: 180
          description: Asientos de la cabina; todos quedan disponibles.
          example: 180
          maximum: 400
          minimum: 1
          type: number
        codigoVuelo:
          description: Dos caracteres de aerolínea y 1 a 4 dígitos. Los dos primeros son el código de la aerolínea.
          example: LA900
          type: string
        destino:
          example: SCL
          type: string
        duracionMinutos:
          example: 300
          maximum: 1500
          minimum: 20
          type: number
        origen:
          description: Código IATA, debe existir en el catálogo de localidades.
          example: BOG
          type: string
        precioBaseUsd:
          description: Tarifa base por adulto en USD, antes de familia e impuestos.
          example: 320
          type: number
        salida:
          description: Salida (UTC), en el futuro.
          example: '2026-11-20T14:30:00.000Z'
          type: string
      required:
        - codigoVuelo
        - aerolinea
        - origen
        - destino
        - salida
        - duracionMinutos
        - precioBaseUsd
      type: object
...
    AdminOrdenViewDto:
      properties:
        _links:
          additionalProperties:
            format: uri
            type: string
          description: HATEOAS links (Richardson Maturity Model Level 3)
          type: object
        bookingId:
          description: >-
            Reserva del núcleo de vuelos tras la emisión: es el `bookingId` de las rutas de equipaje, cambio, cancelación, check-in y pases.
            Solo en la vista del propietario.
          format: uuid
          type: string
        canal:
          type: string
        clienteId:
          description: Cuenta del comprador; null si compró como invitado.
          format: uuid
          nullable: true
          type: string
        comprador:
          enum:
            - cliente
            - invitado
          type: string
        contacto:
          description: Solo en la vista del propietario; la recuperación pública lo enmascara.
          type: object
        creadaEn:
          type: string
        estado:
          enum:
            - PENDIENTE_PAGO
            - PAGO_EN_VERIFICACION
            - PAGADA
            - EMITIDA
            - FALLIDA_COMPENSADA
            - MODIFICADA
            - DEVOLUCION_EN_CURSO
            - REEMBOLSADA
            - EN_VIAJE
            - COMPLETADA
            - EXPIRADA
          type: string
        historial:
          items:
            type: string
          type: array
        itinerarios:
          items:
            $ref: '#/components/schemas/ItinerarioOrdenDto'
          type: array
        mercado:
          type: string
        numeroOrden:
          example: ORD-7K3M9PQ2XA
          type: string
        ordenId:
          format: uuid
          type: string
        pago:
          type: object
        pasajeros:
          items:
            $ref: '#/components/schemas/PasajeroOrdenViewDto'
          type: array
        pnr:
          description: Localizador de la reserva (RN-20).
          nullable: true
          type: string
        total:
          $ref: '#/components/schemas/MontoDto'
      required:
        - ordenId
        - numeroOrden
        - estado
        - mercado
        - total
        - itinerarios
        - pasajeros
        - pago
        - creadaEn
        - historial
        - comprador
        - canal
      type: object
    AdminOrdenesPaginaDto:
      properties:
        items:
          items:
            $ref: '#/components/schemas/AdminOrdenViewDto'
          type: array
        nextCursor:
          type: string
      required:
        - items
      type: object
    AdminReprogramarVueloDto:
      properties:
        motivo:
          example: Cambio de franja operativa
          type: string
        nuevaSalida:
          description: Nueva hora de salida (UTC), en el futuro. La llegada se recalcula conservando la duración.
          example: '2026-11-20T14:30:00.000Z'
          type: string
      required:
        - nuevaSalida
      type: object
...
    AdminVueloViewDto:
      properties:
        aerolinea:
          type: string
        asientosDisponibles:
          type: number
        asientosReservados:
          description: Asientos con número ya reservados (los pasajeros que no eligieron asiento no cuentan).
          type: number
        capacidadTotal:
          type: number
        codigoVuelo:
          example: LA800
          type: string
        destino:
          example: SCL
          type: string
        duracionMinutos:
          type: number
        estado:
          description: Un vuelo cancelado está cerrado a la venta y ya no admite ediciones.
          enum:
            - SCHEDULED
            - CANCELLED
          type: string
        llegada:
          type: string
        origen:
          example: BOG
          type: string
        precioBaseUsd:
          description: Tarifa base por adulto en USD, antes de familia e impuestos.
          type: number
        salida:
          type: string
        vueloId:
          format: uuid
          type: string
      required:
        - vueloId
        - codigoVuelo
        - aerolinea
        - origen
        - destino
        - salida
        - llegada
        - duracionMinutos
        - precioBaseUsd
        - asientosDisponibles
        - capacidadTotal
        - asientosReservados
        - estado
      type: object
    AdminVuelosPaginaDto:
      properties:
        items:
          items:
            $ref: '#/components/schemas/AdminVueloViewDto'
          type: array
        nextCursor:
          type: string
      required:
        - items
      type: object
...
    ProblemDetailsDto:
      properties:
        code:
          description: >-
            The contract's ProblemDetails.code enum, plus
            UNAUTHORIZED/FORBIDDEN/NOT_FOUND/CONFLICT/INTERNAL_ERROR/NOT_IMPLEMENTED/SERVICE_UNAVAILABLE for statuses the contract does not
            define.
          example: SEAT_TAKEN
          type: string
        detail:
          type: string
        invalidParams:
          items:
            $ref: '#/components/schemas/InvalidParamDto'
          type: array
        status:
          example: 409
          type: number
        title:
          type: string
        type:
          example: https://api.booking-hub.com/errors/seat-taken
          type: string
      required:
        - type
        - title
        - status
        - code
      type: object
...
  securitySchemes:
    bearer:
      bearerFormat: JWT
      scheme: bearer
      type: http